requirements.txt
flask>=2.0.0
gunicorn>=20.1.0
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.13
langchain-chroma>=0.0.10
langchain-core>=0.1.15
langchain-text-splitters>=0.0.1
openai>=1.3.0
chromadb>=0.4.18
tiktoken>=0.5.1
pypdf>=3.17.1
python-dotenv>=1.0.0
pandas>=2.0.0
PyMuPDF>=1.23.0
Werkzeug>=2.0.0